requirements.txt
python==3.5.2
Django==2.1
Pillow==5.0.0
django-paypal==0.2.7
pytz==2017.3
